refactor(media): 将媒体处理模块拆分为子模块结构
- 将 `media_processing.rs` 拆分为 `types`、`registry`、`tests` 子模块
- 将 `media_processing_service.rs` 拆分为 `avatar`、`thumbnail`、`resolve`、`shared`、`tests` 子模块
- 新增 vips CLI 处理器支持（头像与缩略图场景）
- 新增 storage-native 处理器支持
- 新增处理器候选解析与优先级路由逻辑
