+++ src/services/media_processing_service/avatar.rs
@@ -1,0 +1,273 @@
+use std::io::Cursor;
+use std::path::PathBuf;
+
+use crate::config::media_processing as media_processing_config;
+use crate::errors::{AsterError, MapAsterErr, Result};
+use crate::runtime::PrimaryAppState;
+use crate::services::profile_service::shared::{
+    AVATAR_SIZE_LG, AVATAR_SIZE_SM, MAX_AVATAR_DECODE_ALLOC,
+};
+use crate::types::MediaProcessorKind;
+use image::imageops::FilterType;
+use image::{DynamicImage, GenericImageView, ImageFormat, ImageReader, Limits};
+
+use super::resolve::resolve_avatar_processor;
+use super::shared::{MediaOperation, ProcessedAvatar};
+
+pub async fn probe_vips_cli_command(command: &str) -> Result<String> {
+    let command = media_processing_config::normalize_vips_command(command)?;
+    if !media_processing_config::command_is_available(&command) {
+        return Err(AsterError::validation_error(format!(
+            "vips_cli command '{command}' is not available"
+        )));
+    }
+
+    tracing::debug!(
+        processor = "vips_cli",
+        command = %command,
+        "starting vips CLI probe"
+    );
+
+    let probe_command = command.clone();
+    let output = tokio::task::spawn_blocking(move || {
+        std::process::Command::new(&probe_command)
+            .arg("--version")
+            .output()
+    })
+    .await
+    .map_aster_err_ctx(
+        "vips CLI probe task panicked",
+        AsterError::thumbnail_generation_failed,
+    )?
+    .map_err(|error| {
+        AsterError::validation_error(format!("spawn vips_cli '{command}': {error}"))
+    })?;
+
+    if !output.status.success() {
+        let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
+        let stdout = String::from_utf8_lossy(&output.stdout).trim().to_string();
+        let detail = if !stderr.is_empty() {
+            stderr
+        } else if !stdout.is_empty() {
+            stdout
+        } else {
+            format!("exit status {}", output.status)
+        };
+        return Err(AsterError::validation_error(format!(
+            "vips_cli probe failed for '{command}': {detail}"
+        )));
+    }
+
+    let detail = String::from_utf8_lossy(&output.stdout).trim().to_string();
+    let detail = if !detail.is_empty() {
+        detail
+    } else {
+        String::from_utf8_lossy(&output.stderr).trim().to_string()
+    };
+
+    tracing::debug!(
+        processor = "vips_cli",
+        command = %command,
+        version = detail.as_str(),
+        "vips CLI probe completed"
+    );
+
+    if detail.is_empty() {
+        Ok(format!("vips_cli command '{command}' is available"))
+    } else {
+        Ok(format!(
+            "vips_cli command '{command}' is available: {detail}"
+        ))
+    }
+}
+
+pub async fn process_avatar_upload(
+    state: &PrimaryAppState,
+    file_name: &str,
+    data: Vec<u8>,
+) -> Result<ProcessedAvatar> {
+    let processor = resolve_avatar_processor(&state.runtime_config, file_name)?;
+    let source_extension = media_processing_config::file_extension(file_name);
+    tracing::debug!(
+        operation = MediaOperation::Avatar.as_str(),
+        processor = processor.kind().as_str(),
+        file_name,
+        source_extension = source_extension.as_deref().unwrap_or(""),
+        source_bytes = data.len(),
+        "processing avatar upload via resolved media processor"
+    );
+
+    match processor.kind() {
+        MediaProcessorKind::Images => {
+            tokio::task::spawn_blocking(move || generate_avatar_variants(data))
+                .await
+                .map_aster_err_ctx(
+                    "avatar processing task panicked",
+                    AsterError::file_upload_failed,
+                )?
+        }
+        MediaProcessorKind::VipsCli => {
+            let command = processor.vips_command().to_string();
+            render_avatar_with_vips_cli(state, file_name, data, &command).await
+        }
+        MediaProcessorKind::StorageNative => Err(AsterError::precondition_failed(
+            "storage-native avatar processing is not supported",
+        )),
+    }
+}
+
+pub(super) fn generate_avatar_variants(data: Vec<u8>) -> Result<ProcessedAvatar> {
+    let mut reader = ImageReader::new(Cursor::new(data))
+        .with_guessed_format()
+        .map_aster_err_ctx("guess avatar format", AsterError::file_type_not_allowed)?;
+
+    let mut limits = Limits::default();
+    limits.max_alloc = Some(MAX_AVATAR_DECODE_ALLOC);
+    reader.limits(limits);
+
+    let img = reader
+        .decode()
+        .map_aster_err_ctx("decode avatar", AsterError::file_type_not_allowed)?;
+
+    let (width, height) = img.dimensions();
+    if width == 0 || height == 0 {
+        return Err(AsterError::validation_error("empty image"));
+    }
+
+    let side = width.min(height);
+    let left = (width - side) / 2;
+    let top = (height - side) / 2;
+    let square = img.crop_imm(left, top, side, side);
+
+    let large = square.resize_exact(AVATAR_SIZE_LG, AVATAR_SIZE_LG, FilterType::Triangle);
+    let small = square.resize_exact(AVATAR_SIZE_SM, AVATAR_SIZE_SM, FilterType::Triangle);
+
+    let large_bytes = encode_avatar_webp(&large)?;
+    let small_bytes = encode_avatar_webp(&small)?;
+
+    Ok(ProcessedAvatar {
+        small_bytes,
+        large_bytes,
+    })
+}
+
+fn encode_avatar_webp(img: &DynamicImage) -> Result<Vec<u8>> {
+    let mut buf = Cursor::new(Vec::new());
+    img.write_to(&mut buf, ImageFormat::WebP)
+        .map_aster_err_ctx("encode avatar webp", AsterError::file_upload_failed)?;
+    Ok(buf.into_inner())
+}
+
+async fn render_avatar_with_vips_cli(
+    state: &PrimaryAppState,
+    file_name: &str,
+    original: Vec<u8>,
+    command: &str,
+) -> Result<ProcessedAvatar> {
+    let temp_root = crate::utils::paths::runtime_temp_dir(&state.config.server.temp_dir);
+    let temp_dir =
+        PathBuf::from(temp_root).join(format!("media-vips-avatar-{}", uuid::Uuid::new_v4()));
+    tokio::fs::create_dir_all(&temp_dir)
+        .await
+        .map_aster_err_ctx(
+            "create avatar vips temp dir",
+            AsterError::storage_driver_error,
+        )?;
+
+    let source_extension =
+        media_processing_config::file_extension(file_name).unwrap_or_else(|| "bin".to_string());
+    let input_path = temp_dir.join(format!("source.{source_extension}"));
+    let small_output_path = temp_dir.join("avatar-512.webp");
+    let large_output_path = temp_dir.join("avatar-1024.webp");
+    tokio::fs::write(&input_path, original)
+        .await
+        .map_aster_err_ctx(
+            "write avatar vips source temp file",
+            AsterError::file_upload_failed,
+        )?;
+
+    let command = command.to_string();
+    let input_arg = input_path.to_string_lossy().to_string();
+    let small_output_arg = small_output_path.to_string_lossy().to_string();
+    let large_output_arg = large_output_path.to_string_lossy().to_string();
+    tracing::debug!(
+        operation = MediaOperation::Avatar.as_str(),
+        processor = MediaProcessorKind::VipsCli.as_str(),
+        command,
+        input_path = input_arg,
+        small_output_path = small_output_arg,
+        large_output_path = large_output_arg,
+        "starting vips CLI avatar render"
+    );
+    tokio::task::spawn_blocking(move || {
+        for (size, output_arg) in [
+            (AVATAR_SIZE_SM, &small_output_arg),
+            (AVATAR_SIZE_LG, &large_output_arg),
+        ] {
+            let output = std::process::Command::new(&command)
+                .arg("thumbnail")
+                .arg(&input_arg)
+                .arg(output_arg)
+                .arg(size.to_string())
+                .arg("--height")
+                .arg(size.to_string())
+                .arg("--size")
+                .arg("both")
+                .arg("--crop")
+                .arg("centre")
+                .output()
+                .map_err(|error| {
+                    AsterError::file_upload_failed(format!(
+                        "spawn avatar vips CLI '{command}': {error}"
+                    ))
+                })?;
+            if !output.status.success() {
+                let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
+                let stdout = String::from_utf8_lossy(&output.stdout).trim().to_string();
+                let detail = if !stderr.is_empty() {
+                    stderr
+                } else if !stdout.is_empty() {
+                    stdout
+                } else {
+                    format!("exit status {}", output.status)
+                };
+                return Err(AsterError::file_upload_failed(format!(
+                    "vips CLI avatar command failed for {size}px output: {detail}"
+                )));
+            }
+        }
+        Ok::<(), AsterError>(())
+    })
+    .await
+    .map_aster_err_ctx(
+        "avatar vips CLI task panicked",
+        AsterError::file_upload_failed,
+    )??;
+
+    let small_bytes = tokio::fs::read(&small_output_path)
+        .await
+        .map_aster_err_ctx(
+            "read avatar vips 512 output",
+            AsterError::file_upload_failed,
+        )?;
+    let large_bytes = tokio::fs::read(&large_output_path)
+        .await
+        .map_aster_err_ctx(
+            "read avatar vips 1024 output",
+            AsterError::file_upload_failed,
+        )?;
+    let _ = tokio::fs::remove_dir_all(&temp_dir).await;
+
+    tracing::debug!(
+        operation = MediaOperation::Avatar.as_str(),
+        processor = MediaProcessorKind::VipsCli.as_str(),
+        small_bytes = small_bytes.len(),
+        large_bytes = large_bytes.len(),
+        "avatar vips CLI render completed"
+    );
+
+    Ok(ProcessedAvatar {
+        small_bytes,
+        large_bytes,
+    })
+}
